+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add cloud / disk as search filters.
 - [Issue 537](https://github.com/aza547/wow-recorder/issues/537) - Space bar can now be used to start/stop the video player.
+- [Issue 522](https://github.com/aza547/wow-recorder/issues/522) - Adds chinese client support for Window and Game capture modes.
 
 ### Fixed
 - [Issue 542](https://github.com/aza547/wow-recorder/issues/542) - Fix a bug with the Devour M+ affix not displaying an image.